Fix STREAMS header not showing in the absence of starred streams.
We wanted to only remove the STARRED header if there are no starred
streams, not the STREAMS header as well.  We do this by switching the
content back to the templates since we're in general avoiding writing
HTML directly in the Javascript since it is bad for e.g. translations.