Change .stream-row borders to #eee.
Change the .stream-row borders to #eee so that we can get rid of the
no-border class. They also look better this way because they are
lighter than the main borders of the modal components which means that
they should be interpreted as soft borders in a component inside the
modal.